debag problem GraphQl

+++ Models/Bases/UserBase.cs
@@ -5,6 +5,6 @@ namespace Models.Bases
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string MiddleName { get; set; }
-        public ushort Age { get; set; }
+        public int Age { get; set; }
     }
 }
